% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfsnormMix.R
\name{dfsnormMix}
\alias{dfsnormMix}
\title{Density function of the a mixture of two skewed-normal distributions}
\usage{
dfsnormMix(x, p, gamma1, mu1, sigma1, gamma2, mu2, sigma2)
}
\arguments{
\item{x}{real: values where the density has to be evaluated.}

\item{gamma1}{positive real: skewness parameter of the first population.}

\item{mu1}{real: location parameter of the first population.}

\item{sigma1}{positive real: scale parameter of the first population.}

\item{gamma2}{positive real: skewness parameter of the second population.}

\item{mu2}{real: location parameter of the second population.}

\item{sigma2}{positive real: scale parameter of the second population.}
}
\value{
Value of the density function of a mixture of two skewed normal
distributions.
}
\description{
This function computes the density of a mixture of two skewed-normal
distributions.
}
\examples{
yd <- dfsnormMix(1,1.3,1,0.5,.8,3,.7)

}
