% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfst.R
\name{dfst}
\alias{dfst}
\title{Density function of the a skewed t distribution}
\usage{
dfst(x, gammap, mu, sigma, nu)
}
\arguments{
\item{x}{real: values where the density has to be evaluated.}

\item{gammap}{real: skewness parameter.}

\item{mu}{real: location parameter.}

\item{sigma}{positive real: scale parameter.}

\item{nu}{real: number of degrees of freedom.}
}
\value{
value of the density function of the skewed t distribution.
}
\description{
This function computes the density of a skewed t distribution.
}
\examples{
yd <- dfst(1,1.3,1,0.5,3)
}
