% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfstmix.R
\name{dfstMix}
\alias{dfstMix}
\title{Density function of the a mixture of two skewed t distributions}
\usage{
dfstMix(x, p, gamma1, mu1, sigma1, nu1, gamma2, mu2, sigma2, nu2)
}
\arguments{
\item{x}{real: values where the density has to be evaluated.}

\item{gamma1}{positive real: skewness parameter of the first population.}

\item{mu1}{real: location parameter of the first population.}

\item{sigma1}{positive real: scale parameter of the first population.}

\item{nu1}{real: number of degrees of freedom of the first population.}

\item{gamma2}{positive real: skewness parameter of the second population.}

\item{mu2}{real: location parameter of the second population.}

\item{sigma2}{positive real: scale parameter of the second population.}

\item{nu2}{real: number of degrees of freedom of the second population.}
}
\value{
Value of the density function of the skewed t distribution. Notice
that, if \eqn{\nu_i=+\infty} (\eqn{i=1,2}), the corresponding component
becomes a skewed normal density.
}
\description{
This function computes the density of a mixture of two skewed t
distributions.
}
\examples{
yd <- dfstMix(1,1.3,1,0.5,3,.8,3,.7,5)

}
