% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{distortion}
\alias{distortion}
\title{Shiller's distortion data}
\format{
A numerical vector with 1857 rows and 1 column.
}
\source{
The data needed for the computation of the S&P 500 distortion
can be downloaded from \url{http://www.econ.yale.edu/~shiller/data.htm}
}
\usage{
distortion
}
\description{
A dataset containing the monthly distortion of the S&P 500 index
from January 1871 until August 2025. The distortion is defined as the
log difference between the real stock market index and its real fundamental
value; see Shiller (2015).
}
\keyword{datasets}
