% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dsnorm.R
\name{dsnorm}
\alias{dsnorm}
\title{Density function of the skewed normal distribution}
\usage{
dsnorm(x, gammap, mu, sigma)
}
\arguments{
\item{x}{real: values where the density has to be evaluated.}

\item{gammap}{real: skewness parameter.}

\item{mu}{real: location parameter.}

\item{sigma}{positive real: scale parameter.}
}
\value{
value of the density function of the skewed normal distribution.
}
\description{
This function computes the density of a skewed normal distribution.
}
\examples{
yd <- dsnorm(1,1.3,1,0.5)
}
