% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/llfst.R
\name{llfst}
\alias{llfst}
\title{Log-likelihood of a skewed t distribution}
\usage{
llfst(x, y)
}
\arguments{
\item{x}{numerical vector (4x1): values of the parameters \eqn{\gamma},
\eqn{\mu}, \eqn{sigma} and \eqn{\nu}.}

\item{y}{numerical vector (nx1): observed data.}
}
\value{
llik real: numerical value of the log-likelihood
function
}
\description{
This function evaluates the log-likelihood of a skewed t distribution.
}
\examples{
y <- rfst(100,1.2,1.3,0.5,3)
x <- c(1.2,1.3,0.5,3)
llik <- llfst(x,y)

}
