% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/llsnormWei.R
\name{llsnormWei}
\alias{llsnormWei}
\title{Weighted log-likelihood of a skewed normal distribution}
\usage{
llsnormWei(x, y, post)
}
\arguments{
\item{x}{numerical vector (3x1): values of the parameters \eqn{\gamma},
\eqn{\mu} and \eqn{\sigma}.}

\item{y}{numerical vector (nx1): observed data.}

\item{post}{numerical vector (nx1) with elements in (0,1): weights
of the observations (in the EM algorithm, posterior probabilities).}
}
\value{
Numerical value of the log-likelihood
function
}
\description{
This function evaluates the log-likelihood of a skewed normal distribution
with weighted observations.
}
\examples{
y <- rfst(100,1.2,1.3,0.5,300)
x <- c(1.2,1.3,0.5,3)
post <- runif(100)
llik <- llsnormWei(x,y,post)

}
