% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pfsnorm.R
\name{pfsnorm}
\alias{pfsnorm}
\title{Cumulative distribution function of the skewed-normal distribution}
\usage{
pfsnorm(x, gammap, mu, sigma)
}
\arguments{
\item{x}{real vector: values where the cdf has to be evaluated.}

\item{gammap}{positive real: skewness parameter.}

\item{mu}{real: location parameter.}

\item{sigma}{positive real: scale parameter.}
}
\value{
Value of the cdf of the skewed t distribution.
}
\description{
This function evaluates the cdf of the skewed-normal distribution
for a vectorial input
}
\examples{
yd <- pfsnorm(1,1.3,1,0.5)
}
