% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pfsnormMix.R
\name{pfsnormMix}
\alias{pfsnormMix}
\title{Cumulative distribution function of a two-population mixture of
skewed-normal distributions}
\usage{
pfsnormMix(x, p, gamma1, mu1, sigma1, gamma2, mu2, sigma2)
}
\arguments{
\item{x}{real vector: values where the cdf has to be evaluated.}

\item{p}{real 0<=p<=1: mixing weight.}

\item{gamma1}{positive real: skewness parameter of the first population.}

\item{mu1}{real: location parameter of the first population.}

\item{sigma1}{positive real: scale parameter of the first population.}

\item{gamma2}{positive real: skewness parameter of the second population.}

\item{mu2}{real: location parameter of the second population.}

\item{sigma2}{positive real: scale parameter of the second population.}
}
\value{
Value of the cdf of a two-population mixture of
skewed-normal distributions.
}
\description{
This function evaluates the cdf of a two-population mixture of
skewed-normal distributions for a vectorial input
}
\examples{
yd <- pfsnormMix(1,0.5,1.3,-1,0.5,.7,2,1)
}
