% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pfst.R
\name{pfst}
\alias{pfst}
\title{Cumulative distribution function of the skewed Student t distribution}
\usage{
pfst(x, gammap, mu, sigma, nu)
}
\arguments{
\item{x}{real vector: values where the cdf has to be evaluated.}

\item{gammap}{positive real: skewness parameter.}

\item{mu}{real: location parameter.}

\item{sigma}{positive real: scale parameter.}

\item{nu}{positive real: number of degrees of freedom.}
}
\value{
value of the cdf of the skewed t distribution.
}
\description{
This function evaluates the cdf of the skewed Student t distribution
for a vectorial input
}
\examples{
yd <- pfst(1,1.3,1,0.5,3)
}
