% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pfstMix.R
\name{pfstMix}
\alias{pfstMix}
\title{Cumulative distribution function of a two-population mixture of
skewed t distributions}
\usage{
pfstMix(x, p, gamma1, mu1, sigma1, nu1, gamma2, mu2, sigma2, nu2)
}
\arguments{
\item{x}{real vector: values where the cdf has to be evaluated.}

\item{p}{real 0<=p<=1: mixing weight.}

\item{gamma1}{positive real: skewness parameter of the first population.}

\item{mu1}{real: location parameter of the first population.}

\item{sigma1}{positive real: scale parameter of the first population.}

\item{nu1}{positive real: number of degrees of freedomr of the first
population.}

\item{gamma2}{positive real: skewness parameter of the second population.}

\item{mu2}{real: location parameter of the second population.}

\item{sigma2}{positive real: scale parameter of the second population.}

\item{nu2}{positive real: number of degrees of freedom of the second
population.}
}
\value{
value of the cdf of of a two-population mixture of
skewed t distributions.
}
\description{
This function evaluates the cdf of a two-population mixture of
skewed t distributions for a vectorial input
}
\examples{
yd <- pfst(1,1.3,1,0.5,3)
}
