% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qfsnorm.R
\name{qfsnorm}
\alias{qfsnorm}
\title{Quantile function of the skewed-normal distribution}
\usage{
qfsnorm(w, gammap, mu, sigma)
}
\arguments{
\item{w}{real, 0<=w<=1: quantile levels.}

\item{gammap}{positive real: skewness parameter.}

\item{mu}{real: location parameter.}

\item{sigma}{positive real: scale parameter.}
}
\value{
Value of the quantile function of the skewed t distribution.
}
\description{
This function evaluates the quantile function of the skewed
Student t distribution
}
\examples{
yd <- qfsnorm(.1,1.2,1,0.5)
}
