% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qfsnormMix.R
\name{qfsnormMix}
\alias{qfsnormMix}
\title{Quantile function of a mixture of skewed normal t distributions}
\usage{
qfsnormMix(w, p, gamma1, mu1, sigma1, gamma2, mu2, sigma2)
}
\arguments{
\item{w}{real, 0<=w<=1: quantile levels.}

\item{p}{real 0<=p<=1: mixing weight.}

\item{gamma1}{positive real: skewness parameter of the first population.}

\item{mu1}{real: location parameter of the first population.}

\item{sigma1}{positive real: scale parameter of the first population.}

\item{gamma2}{positive real: skewness parameter of the second population.}

\item{mu2}{real: location parameter of the second population.}

\item{sigma2}{positive real: scale parameter of the second population.}
}
\value{
value of the quantile function of a mixture of skewed-normal
distributions.
}
\description{
This function evaluates the Quantile function of a mixture of skewed
normal distributions
}
\examples{
yd <- qfsnormMix(.9,.1,.8,-1,0.5,1.3,2,1)
}
