% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qfst.R
\name{qfst}
\alias{qfst}
\title{Quantile function of the skewed Student t distribution}
\usage{
qfst(w, gammap, mu, sigma, nu)
}
\arguments{
\item{w}{real, 0<=w<=1: quantile levels.}

\item{gammap}{positive real: skewness parameter.}

\item{mu}{real: location parameter.}

\item{sigma}{positive real: scale parameter.}

\item{nu}{positive real: number of degrees of freedom.}
}
\value{
value of the quantile function of the skewed t distribution.
}
\description{
This function evaluates the Quantile function of the skewed
Student t distribution
}
\examples{
yd <- qfst(.1,1.2,1,0.5,3)
}
