% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qfstMix.R
\name{qfstMix}
\alias{qfstMix}
\title{Quantile function of a mixture of skewed Student t distributions}
\usage{
qfstMix(w, p, gamma1, mu1, sigma1, nu1, gamma2, mu2, sigma2, nu2)
}
\arguments{
\item{w}{real, 0<=w<=1: quantile levels.}

\item{p}{real 0<=p<=1: mixing weight.}

\item{gamma1}{positive real: skewness parameter of the first population.}

\item{mu1}{real: location parameter of the first population.}

\item{sigma1}{positive real: scale parameter of the first population.}

\item{nu1}{positive real: number of degrees of freedom of the first
population.}

\item{gamma2}{positive real: skewness parameter of the second population.}

\item{mu2}{real: location parameter of the second population.}

\item{sigma2}{positive real: scale parameter of the second population.}

\item{nu2}{positive real: number of degrees of freedom of the second
population.}
}
\value{
value of the quantile function of a mixture of skewed t
distributions. Notice that, if \eqn{\nu_i=+\infty} (\eqn{i=1,2}), the
corresponding component becomes a skewed normal density.
}
\description{
This function evaluates the Quantile function of a mixture of skewed
t distributions
}
\examples{
yd <- qfst(.1,1.2,1,0.5,3)
}
