% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rfsnormMix.R
\name{rfsnormMix}
\alias{rfsnormMix}
\title{Random number simulation from a two-population mixture of skewed-normal
distributions}
\usage{
rfsnormMix(n, p, gamma1, mu1, sigma1, gamma2, mu2, sigma2)
}
\arguments{
\item{n}{positive integer: number of simulated random numbers.}

\item{p}{real 0<=p<=1: mixing weight.}

\item{gamma1}{positive real: skewness parameter of the first population.}

\item{mu1}{real: location parameter of the first population.}

\item{sigma1}{positive real: scale parameter of the first population.}

\item{gamma2}{positive real: skewness parameter of the second population.}

\item{mu2}{real: location parameter of the second population.}

\item{sigma2}{positive real: scale parameter of the second population.}
}
\value{
A vector containing n iid random numbers from a mixture of skewed-
normal distributions.
}
\description{
This function simulates random numbers from a two-population
mixture of skewed-normal distributions
}
\examples{
ySim <- rfsnormMix(100,1.2,1.3,0.5,.8,.5,1)
}
