% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rfst.R
\name{rfst}
\alias{rfst}
\title{Random number simulation from a skewed t distribution}
\usage{
rfst(n, gammap, mu, sigma, nu)
}
\arguments{
\item{n}{positive integer: number of simulated random numbers.}

\item{gammap}{positive real: skewness parameter.}

\item{mu}{real: location parameter.}

\item{sigma}{positive real: scale parameter.}

\item{nu}{positive real: number of degrees of freedom.}
}
\value{
A vector containing n iid random numbers from the skewed t distribution.
}
\description{
This function simulates random numbers from a skewed t distribution
}
\examples{
ySim <- rfst(100,1.2,1.3,0.5,3)
}
