% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snormMixEM01.R
\name{snormMixEM01}
\alias{snormMixEM01}
\title{Skewed-normal mixture estimation via EM}
\usage{
snormMixEM01(Y, p1, mu1, sigma1, gamma2, mu2, sigma2, maxiter)
}
\arguments{
\item{maxiter}{positive integer: maximum number of iterations of the EM algorithm.}

\item{x0}{numerical vector (7x1): initial values of the parameters \emph{p},
\eqn{\mu_1,} \eqn{\sigma_1}
\eqn{\gamma_2}, \eqn{\mu_2}, \eqn{\sigma_2}.}

\item{y}{vector: observed data.}
}
\value{
A list with the following elements is returned:
"p" = estimated value of p,
"post" = posterior probabilities of all observations,
"mu1" = estimated value of \eqn{\mu_1},
"sigma1 " = estimated value of \eqn{\sigma_1},
"gamma2" = estimated value of \eqn{\gamma_2},
"mu2" = estimated value of \eqn{\mu_2},
"sigma2 " = estimated value of \eqn{\sigma_2},
"nit" = number of iterations.
}
\description{
This function estimates a of a two-population mixture of
skewed t distributions by means of the EM algorithm.
}
\examples{
y <- rfstMix(100,.5,1,1.3,0.5,.8,1,1)
x0 <- c(.6,1.3,1,0.8,1.2,.7,.9)
res <- stMixEM(y,x0)

}
