% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stMixEMSwitch.R
\name{stMixEMSwitch}
\alias{stMixEMSwitch}
\title{Mixture estimation via EM}
\usage{
stMixEMSwitch(
  Y,
  p1,
  gamma1,
  mu1,
  sigma1,
  nu1,
  gamma2,
  mu2,
  sigma2,
  nu2,
  nustar,
  maxiter
)
}
\arguments{
\item{maxiter}{positive integer: maximum number of iterations of the EM algorithm.}

\item{x0}{numerical vector (9x1): initial values of the parameters \emph{p},
\eqn{\gamma_1}, \eqn{\mu_1,} \eqn{\sigma_1}, \eqn{\nu_1},
\eqn{\gamma_2}, \eqn{\mu_2}, \eqn{\sigma_2}, \eqn{\nu_2}.}

\item{y}{vector: observed data.}
}
\value{
A list with the following elements is returned:
"p" = estimated value of p,
"post" = posterior probabilities of all observations,
"gamma1" = estimated value of \eqn{\gamma_1},
"mu1" = estimated value of \eqn{\mu_1},
"sigma1 " = estimated value of \eqn{\sigma_1},
"nu1" = estimated value of \eqn{\nu_1},
"gamma2" = estimated value of \eqn{\gamma_2},
"mu2" = estimated value of \eqn{\mu_2},
"sigma2 " = estimated value of \eqn{\sigma_2},
"nu2" = estimated value of \eqn{\nu_2},
"nit" = number of iterations.
}
\description{
This function estimates a of a two-population mixture of
skewed t distributions by means of the EM algorithm.
}
\examples{
y <- rfstMix(100,.5,1.2,1.3,0.5,3,.8,1,1,7)
x <- c(.6,1.3,1,0.8,3,1.2,.7,.9,9)
res <- stMixEM(y,.6,1.3,1,0.8,3,1.2,.7,.9,9,30,1000)

}
